+++ config/routes.rb
@@ -1,23 +1,11 @@
 Rails.application.routes.draw do
-  # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
-
-  # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
-  # Can be used by load balancers and uptime monitors to verify that the app is live.
-  get "up" => "rails/health#show", as: :rails_health_check
-
-  # Render dynamic PWA files from app/views/pwa/* (remember to link manifest in application.html.erb)
-  # get "manifest" => "rails/pwa#manifest", as: :pwa_manifest
-  # get "service-worker" => "rails/pwa#service_worker", as: :pwa_service_worker
-
-  # Defines the root path route ("/")
-  # root "posts#index"
-
-  post "users", to: "users#create"
-  post "users/login", to: "users#login"
-
-  post "tasks", to: "tasks#create"
-  get "tasks", to: "tasks#index"
-  get "tasks/:identity", to: "tasks#show"
-  delete "tasks/:identity", to: "tasks#destroy"
-  put "tasks/:identity", to: "tasks#update"
+  namespace :v1 do
+    resources :users, only: [:create] do
+      collection do
+        post "login", to: "users#login"
+      end
+    end
+  
+    resources :tasks, only: [:create, :index, :show, :update, :destroy], param: :identity
+  end
 end
